improved merge ui in dishGridView
- update row data locally after mutate without refetch
- Prominently show amount of selected dishes in merge button.
The Checkbox in the table header can be misleading as it only refers to
the currently visible elements. Thus,
when you selected an item and then change to a page
where the item is not visible(
e.g. change the filter
to search for a specific item),
the checkbox would be empty
suggesting that no item is selected,
which is misleading
- added clear selection button
- don't reset grid after merging